refactor: rename window module
